requirements.txt
asgiref==3.8.1
boto3==1.34.108
botocore==1.34.108
Django==5.0.6
django-cors-headers==4.3.1
django-storages==1.14.3
djangorestframework==3.15.1
gunicorn==22.0.0
jmespath==1.0.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
s3transfer==0.10.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
supervisor==4.2.5
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0